+++ jest.e2e.setup.js
@@ -1,3 +1,4 @@
 /* global jest */
+// import "regenerator-runtime/runtime";
 
-jest.setTimeout(10000);
+jest.setTimeout(20000);