+++ src/plugins/files-loader/FilesLoader.js
@@ -13,18 +13,18 @@ const path = require("path");
 const globule = require("globule");
 const watch = require("node-watch");
 const fsExtra = require("fs-extra");
-
 const { map, debounce, flatten } = require("lodash");
 
-const { mocksFileToUse } = require("./helpers");
+const { mocksFileToUse, babelRegisterDefaultOptions, getFilesGlobule } = require("./helpers");
 const { createMocksFolder } = require("../../support/scaffold");
 
 const PLUGIN_NAME = "@mocks-server/core/plugin-files-loader";
+const BABEL_REGISTER_OPTION = "babelRegister";
+const BABEL_REGISTER_OPTIONS_OPTION = "babelRegisterOptions";
 const PATH_OPTION = "path";
 const WATCH_OPTION = "watch";
 const DEFAULT_PATH = "mocks";
 const ROUTES_FOLDER = "routes";
-const MOCKS_FILE = "mocks";
 
 class FilesLoaderBase {
   constructor(core, methods, extraOptions = {}) {
@@ -76,6 +76,11 @@ class FilesLoaderBase {
     }
   }
 
+  _readFile(filePath) {
+    const content = this._require(filePath);
+    return content.default || content;
+  }
+
   _cleanRequireCacheFolder() {
     map(this._cache(), (cacheData, filePath) => {
       if (filePath.indexOf(this._path) === 0) {
@@ -125,6 +130,14 @@ class FilesLoaderBase {
     const resolvedFolder = this._resolveFolder(pathName);
     this._path = this._ensureFolder(resolvedFolder);
     this._tracer.info(`Loading files from folder ${this._path}`);
+    if (!!this._core.lowLevelConfig[BABEL_REGISTER_OPTION]) {
+      this._require("@babel/register")(
+        babelRegisterDefaultOptions(
+          resolvedFolder,
+          this._core.lowLevelConfig[BABEL_REGISTER_OPTIONS_OPTION]
+        )
+      );
+    }
     this._cleanRequireCacheFolder();
     this._loadRoutesFiles();
     this._loadMocksFile();
@@ -134,7 +147,10 @@ class FilesLoaderBase {
     const routesPath = path.resolve(this._path, ROUTES_FOLDER);
     try {
       const routeFiles = globule.find({
-        src: ["**/*.js", "**/*.json"],
+        src: getFilesGlobule(
+          this._core.lowLevelConfig[BABEL_REGISTER_OPTION],
+          this._core.lowLevelConfig[BABEL_REGISTER_OPTIONS_OPTION]
+        ),
         srcBase: routesPath,
         prefixBase: true,
       });
@@ -142,7 +158,7 @@ class FilesLoaderBase {
       const routes = flatten(
         routeFiles.map((filePath) => {
           // TODO, validate basic routes structure, add warning for not valid routes
-          return this._require(filePath);
+          return this._readFile(filePath);
         })
       );
       this._loadRoutes(routes);
@@ -155,12 +171,14 @@ class FilesLoaderBase {
   }
 
   _loadMocksFile() {
-    const mocksFileJs = path.resolve(this._path, `${MOCKS_FILE}.js`);
-    const mocksFileJson = path.resolve(this._path, `${MOCKS_FILE}.json`);
-    let mocksFile = mocksFileToUse(mocksFileJs, mocksFileJson);
+    let mocksFile = mocksFileToUse(
+      this._path,
+      this._core.lowLevelConfig[BABEL_REGISTER_OPTION],
+      this._core.lowLevelConfig[BABEL_REGISTER_OPTIONS_OPTION]
+    );
     if (mocksFile) {
       try {
-        const mocks = this._require(mocksFile);
+        const mocks = this._readFile(mocksFile);
         // TODO, validate mocks, add warning for not valid mocks
         this._loadMocks(mocks);
         this._tracer.silly(`Loaded mocks from file ${mocksFile}`);